Removing pydase/utils/logging.py from pytest coverage report
Changing the environment variable "ENVIRONMENT" does not change the
confz configuration. Additionally, the setup_logging file is removing
any loggers (line 61). With this, caplog with loguru doesn't work as we
are already using some sort of hack to get the loguru logs with pytest.
We can therefore not test if the setup_logging function works as
expected.